+++ example/lib/startback_todo/database.rb
@@ -1,0 +1,16 @@
+module StartbackTodo
+  class Database
+
+    def initialize
+      @todos ||= [
+        {id: 1, description: "Write more code"}
+      ]
+    end
+    attr_reader :todos
+
+    def transaction
+      yield
+    end
+
+  end
+end