+++ example/lib/startback_todo/operation.rb
@@ -1,0 +1,5 @@
+module StartbackTodo
+  class Operation < Startback::Operation
+  end
+end
+require_relative "operation/create_todo"