+++ example/lib/startback_todo/operation/create_todo.rb
@@ -1,0 +1,13 @@
+module StartbackTodo
+  class CreateTodo < Operation
+
+    def initialize(todo)
+      @todo = todo
+    end
+
+    def call
+      db.todos << @todo
+    end
+
+  end
+end