+++ lib/startback/support.rb
@@ -15,3 +15,4 @@ module Startback
   end # module Support
 end # module Startback
 require_relative 'support/logger'
+require_relative 'support/operation_runner'